Extract LoaderSpy class to a separate file
