+++ EssentialFeediOSTests/Controllers/Helpers/FeedViewControllerTests+LoaderSpy.swift
@@ -1,0 +1,67 @@
+//
+//  FeedViewControllerTests+LoaderSpy.swift
+//  EssentialFeediOSTests
+//
+//  Created by Jair Moreno Gaspar on 31/12/20.
+//  Copyright © 2020 Jair Moreno G. All rights reserved.
+//
+
+import Foundation
+import EssentialFeediOS
+import EssentialFeed
+
+class LoaderSpy: FeedLoader, FeedImageDataLoader {
+    // MARK: - Feed loader
+    private(set) var feedRequests = [(FeedLoader.Result) -> Void]()
+    
+    var loadFeedCallCount: Int {
+        feedRequests.count
+    }
+    
+    func load(completion: @escaping (FeedLoader.Result) -> Void) {
+        feedRequests.append(completion)
+    }
+    
+    func completeFeedLoading(with feed: [FeedImage] = [], at index: Int = 0) {
+        feedRequests[index](.success(feed))
+    }
+    
+    func completeFeedLoadingWithError(at index: Int) {
+        let error = NSError(domain: "an error", code: 0)
+        feedRequests[index](.failure(error))
+    }
+    
+    // MARK: - FeedImageDataLoader
+    
+    private struct TaskSpy: FeedImageDataLoaderTask {
+        
+        let cancelCallBack: () -> Void
+        
+        func cancel() {
+            cancelCallBack()
+        }
+    }
+    
+    var loadedImageURLs: [URL] {
+        imageRequests.map { $0.url }
+    }
+    
+    private var imageRequests = [(url: URL, completion: (FeedImageDataLoader.Result) -> Void)]()
+    private(set) var cancelledImageURLs = [URL]()
+    
+    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        imageRequests.append((url, completion))
+        return TaskSpy { [weak self] in
+            self?.cancelledImageURLs.append(url)
+        }
+    }
+    
+    func completeImageLoadingWithError(at index: Int = 0) {
+        let error = NSError(domain: "an error", code: 0)
+        imageRequests[index].completion(.failure(error))
+    }
+    
+    func completeImageLoading(with imageData: Data = Data(), at index: Int = 0) {
+        imageRequests[index].completion(.success(imageData))
+    }
+}
